+++ OpenXmlPowerToolsExamples/DocumentAssembler02/DocumentAssembler02.cs
@@ -8,8 +8,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
-using DocumentFormat.OpenXml.Packaging;
-using Clippit;
+using Clippit.Word;
 
 namespace Clippit
 {