+++ OpenXmlPowerToolsExamples/DocumentBuilder01/DocumentBuilder01.cs
@@ -9,6 +9,8 @@ using System.Text;
 using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Validation;
 using Clippit;
+using Clippit.Word;
+
 
 namespace DocumentBuilderExample
 {
@@ -23,10 +25,10 @@ namespace DocumentBuilderExample
             string source1 = "../../Source1.docx";
             string source2 = "../../Source2.docx";
             string source3 = "../../Source3.docx";
-            List<Source> sources = null;
+            List<ISource> sources = null;
 
             // Create new document from 10 paragraphs starting at paragraph 5 of Source1.docx
-            sources = new List<Source>()
+            sources = new List<ISource>()
             {
                 new Source(new WmlDocument(source1), 5, 10, true),
             };
@@ -34,7 +36,7 @@ namespace DocumentBuilderExample
 
             // Create new document from paragraph 1, and paragraphs 5 through end of Source3.docx.
             // This effectively 'deletes' paragraphs 2-4
-            sources = new List<Source>()
+            sources = new List<ISource>()
             {
                 new Source(new WmlDocument(source3), 0, 1, false),
                 new Source(new WmlDocument(source3), 4, false),
@@ -43,7 +45,7 @@ namespace DocumentBuilderExample
 
             // Create a new document that consists of the entirety of Source1.docx and Source2.docx.  Use
             // the section information (headings and footers) from source1.
-            sources = new List<Source>()
+            sources = new List<ISource>()
             {
                 new Source(new WmlDocument(source1), true),
                 new Source(new WmlDocument(source2), false),
@@ -52,7 +54,7 @@ namespace DocumentBuilderExample
 
             // Create a new document that consists of the entirety of Source1.docx and Source2.docx.  Use
             // the section information (headings and footers) from source2.
-            sources = new List<Source>()
+            sources = new List<ISource>()
             {
                 new Source(new WmlDocument(source1), false),
                 new Source(new WmlDocument(source2), true),
@@ -62,7 +64,7 @@ namespace DocumentBuilderExample
             // Create a new document that consists of the first 5 paragraphs of Source1.docx and the first
             // five paragraphs of Source2.docx.  This example returns a new WmlDocument, when you then can
             // serialize to a SharePoint document library, or use in some other interesting scenario.
-            sources = new List<Source>()
+            sources = new List<ISource>()
             {
                 new Source(new WmlDocument(source1), 0, 5, false),
                 new Source(new WmlDocument(source2), 0, 5, true),