+++ OpenXmlPowerToolsExamples/DocumentBuilder02/DocumentBuilder02.cs
@@ -9,6 +9,8 @@ using System.Text;
 using System.Xml.Linq;
 using DocumentFormat.OpenXml.Packaging;
 using Clippit;
+using Clippit.Word;
+
 
 // Insert content into a document
 // Delete content from a document
@@ -31,9 +33,9 @@ class DocumentBuilderExample02
         tempDi.Create();
 
         // Insert an abstract and author biography into a white paper.
-        List<Source> sources = null;
+        List<ISource> sources = null;
 
-        sources = new List<Source>()
+        sources = new List<ISource>()
         {
             new Source(new WmlDocument("../../WhitePaper.docx"), 0, 1, true),
             new Source(new WmlDocument("../../Abstract.docx"), false),
@@ -66,6 +68,7 @@ class DocumentBuilderExample02
                 .Select(g => new Source(
                     new WmlDocument("../../Notes.docx"), g.First().Index,
                         g.Last().Index - g.First().Index + 1, true))
+                .Cast<ISource>()
                 .ToList();
         }
         DocumentBuilder.BuildDocument(sources, Path.Combine(tempDi.FullName, "NewNotes.docx"));
@@ -116,7 +119,7 @@ class DocumentBuilderExample02
         foreach (var doc in documentList)
         {
             string fileName = String.Format("Section{0:000}.docx", doc.DocumentNumber);
-            List<Source> documentSource = new List<Source> {
+            List<ISource> documentSource = new List<ISource> {
                 new Source(new WmlDocument("../../Spec.docx"), doc.Start, doc.Count, true)
             };
             DocumentBuilder.BuildDocument(documentSource, Path.Combine(tempDi.FullName, fileName));
@@ -126,6 +129,7 @@ class DocumentBuilderExample02
         sources = tempDi
             .GetFiles("Section*.docx")
             .Select(d => new Source(new WmlDocument(d.FullName), true))
+            .Cast<ISource>()
             .ToList();
         DocumentBuilder.BuildDocument(sources, Path.Combine(tempDi.FullName, "ReassembledSpec.docx"));
     }