+++ OpenXmlPowerToolsExamples/DocumentBuilder03/DocumentBuilder03.cs
@@ -10,6 +10,8 @@ using System.Xml.Linq;
 using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Validation;
 using Clippit;
+using Clippit.Word;
+
 
 class Program
 {
@@ -45,7 +47,7 @@ class Program
         }
 
         string outFileName = Path.Combine(tempDi.FullName, "Out.docx");
-        List<Source> sources = new List<Source>()
+        List<ISource> sources = new List<ISource>()
             {
                 new Source(doc1, true),
                 new Source(new WmlDocument(@"..\..\Insert-01.docx"), "Liz"),