+++ OpenXmlPowerToolsExamples/DocumentBuilder04/DocumentBuilder04.cs
@@ -10,6 +10,7 @@ using System.Xml;
 using System.Xml.Linq;
 using DocumentFormat.OpenXml.Packaging;
 using Clippit;
+using Clippit.Word;
 
 namespace ExampleDocumentBuilder04
 {
@@ -104,7 +105,7 @@ namespace ExampleDocumentBuilder04
                             else
                                 return new Source(g.Document, false);
                         }
-                    ).ToList();
+                    ).Cast<ISource>().ToList();
 
                 DocumentBuilder.BuildDocument(sources, Path.Combine(tempDi.FullName, "solar-system-new.docx"));
             }