+++ OpenXmlPowerToolsExamples/HtmlToWmlConverter02/HtmlToWmlConverter02.cs
@@ -2,17 +2,15 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.Generic;
 using System.Drawing.Imaging;
 using System.IO;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.Linq;
 using DocumentFormat.OpenXml.Packaging;
-using Clippit;
-using Clippit.HtmlToWml;
+using Clippit.Word;
+
 
 namespace Clippit
 {